refactor layout and display rules to match project requirements
